move to sep module, update test
