+++ packages/next/src/server/base-server.ts
@@ -175,7 +175,7 @@ import type { RouteModule } from './route-modules/route-module'
 import { FallbackMode, parseFallbackField } from '../lib/fallback'
 import { toResponseCacheEntry } from './response-cache/utils'
 import { scheduleOnNextTick } from '../lib/scheduler'
-import { nextConfigNormalizer } from './config-shared'
+import { nextConfigNormalizer } from './serializable-config'
 import { shouldServeStreamingMetadata } from './lib/streaming-metadata'
 
 export type FindComponentsResult = {
